Fix sendFile callback to only call on error
The callback is named Errback as the type, but it is called no matter
if there is an error or not. This fixes an issue where
`ERR_HTTP_HEADERS_SENT` is being thrown for correctly resolved files
because the file is found, sent, and then the 404.html file was
attempted to be sent directly after.

These changes only render a 404 file if the path could not be found.